MMC: JZ4740: Read status register before clearing irq flags
The JZ4750 and JZ4760 have additional bits in the irq register for the different
error conditions. If the bit in the irq register is cleared the matching bit in
the status register is cleared as well, so the status register has to be read
before those bits are cleared, otherwise error conditions are undetected.